requirements.txt
streamlit
Pillow
numpy
joblib
tensorflow-cpu
opencv-python
scikit-learn
streamlit-webrtc